requirements.txt
RPi.Gpio
Adafruit_DHT
time
requests
os